+++ packages/react-router-dev/cli/commands.ts
@@ -1,5 +1,5 @@
 import { existsSync } from "node:fs";
-import { readFile, writeFile } from "node:fs/promises";
+import { readFile, writeFile, copyFile } from "node:fs/promises";
 import { createRequire } from "node:module";
 import * as path from "node:path";
 import exitHook from "exit-hook";
@@ -77,6 +77,7 @@ export async function dev(root?: string, options: ViteDevOptions = {}) {
 let clientEntries = ["entry.client.tsx", "entry.client.js", "entry.client.jsx"];
 let serverEntries = ["entry.server.tsx", "entry.server.js", "entry.server.jsx"];
 let entries = ["entry.client", "entry.server"];
+let rscEntries = ["entry.client", "entry.rsc", "entry.ssr"];
 
 let conjunctionListFormat = new Intl.ListFormat("en", {
   style: "long",
@@ -94,6 +95,10 @@ export async function generateEntry(
 ) {
   rootDirectory = resolveRootDirectory(rootDirectory, flags);
 
+  let configDir = "defaults";
+  let entriesToUse = entries;
+  let isRsc = false;
+
   if (
     await hasReactRouterRscPlugin({
       root: rootDirectory,
@@ -103,12 +108,16 @@ export async function generateEntry(
       },
     })
   ) {
-    console.error(
-      colors.red(
-        `The reveal command is currently not supported in RSC Framework Mode.`,
-      ),
-    );
-    process.exit(1);
+    if (!entry) {
+      await generateEntry("entry.client", rootDirectory, flags);
+      await generateEntry("entry.rsc", rootDirectory, flags);
+      await generateEntry("entry.ssr", rootDirectory, flags);
+      return;
+    }
+
+    configDir = "default-rsc-entries";
+    entriesToUse = rscEntries;
+    isRsc = true;
   }
 
   // if no entry passed, attempt to create both
@@ -130,8 +139,8 @@ export async function generateEntry(
 
   let appDirectory = configResult.value.appDirectory;
 
-  if (!entries.includes(entry)) {
-    let entriesArray = Array.from(entries);
+  if (!entriesToUse.includes(entry)) {
+    let entriesArray = Array.from(entriesToUse);
     let list = conjunctionListFormat.format(entriesArray);
 
     console.error(
@@ -140,48 +149,70 @@ export async function generateEntry(
     return;
   }
 
-  // TODO(v8): Remove - only required for Node 20.18 and below
-  let { readPackageJSON } = await import("pkg-types");
-  let pkgJson = await readPackageJSON(rootDirectory);
-  let deps = pkgJson.dependencies ?? {};
-
-  if (!deps["@react-router/node"]) {
-    console.error(colors.red(`No default server entry detected.`));
-    return;
-  }
-
   let defaultsDirectory = path.resolve(
     path.dirname(nodeRequire.resolve("@react-router/dev/package.json")),
     "dist",
     "config",
-    "defaults",
-  );
-  let defaultEntryClient = path.resolve(defaultsDirectory, "entry.client.tsx");
-
-  let defaultEntryServer = path.resolve(
-    defaultsDirectory,
-    `entry.server.node.tsx`,
+    configDir,
   );
 
-  let isServerEntry = entry === "entry.server";
+  let outputFile: string;
+  if (isRsc) {
+    let defaultEntry = path.resolve(defaultsDirectory, `${entry}.tsx`);
+    outputFile = path.resolve(appDirectory, `${entry}.tsx`);
 
-  let contents = isServerEntry
-    ? await createServerEntry(rootDirectory, appDirectory, defaultEntryServer)
-    : await createClientEntry(rootDirectory, appDirectory, defaultEntryClient);
+    if (existsSync(outputFile)) {
+      let relative = path.relative(rootDirectory, outputFile);
+      console.error(colors.red(`Entry file ${relative} already exists.`));
+      return;
+    }
 
-  let useTypeScript = flags.typescript ?? true;
-  let outputExtension = useTypeScript ? "tsx" : "jsx";
-  let outputEntry = `${entry}.${outputExtension}`;
-  let outputFile = path.resolve(appDirectory, outputEntry);
-
-  if (!useTypeScript) {
-    let javascript = await convertFileToJS(contents, {
-      cwd: rootDirectory,
-      filename: isServerEntry ? defaultEntryServer : defaultEntryClient,
-    });
-    await writeFile(outputFile, javascript, "utf-8");
+    await copyFile(defaultEntry, outputFile);
   } else {
-    await writeFile(outputFile, contents, "utf-8");
+    // TODO(v8): Remove - only required for Node 20.18 and below
+    let { readPackageJSON } = await import("pkg-types");
+    let pkgJson = await readPackageJSON(rootDirectory);
+    let deps = pkgJson.dependencies ?? {};
+
+    if (!deps["@react-router/node"]) {
+      console.error(colors.red(`No default server entry detected.`));
+      return;
+    }
+
+    let defaultEntryClient = path.resolve(
+      defaultsDirectory,
+      "entry.client.tsx",
+    );
+
+    let defaultEntryServer = path.resolve(
+      defaultsDirectory,
+      `entry.server.node.tsx`,
+    );
+
+    let isServerEntry = entry === "entry.server";
+
+    let contents = isServerEntry
+      ? await createServerEntry(rootDirectory, appDirectory, defaultEntryServer)
+      : await createClientEntry(
+          rootDirectory,
+          appDirectory,
+          defaultEntryClient,
+        );
+
+    let useTypeScript = flags.typescript ?? true;
+    let outputExtension = useTypeScript ? "tsx" : "jsx";
+    let outputEntry = `${entry}.${outputExtension}`;
+    outputFile = path.resolve(appDirectory, outputEntry);
+
+    if (!useTypeScript) {
+      let javascript = await convertFileToJS(contents, {
+        cwd: rootDirectory,
+        filename: isServerEntry ? defaultEntryServer : defaultEntryClient,
+      });
+      await writeFile(outputFile, javascript, "utf-8");
+    } else {
+      await writeFile(outputFile, contents, "utf-8");
+    }
   }
 
   console.log(