+++ playground/split-route-modules-spa/vite.config.ts
@@ -1,7 +1,7 @@
 import { reactRouter } from "@react-router/dev/vite";
 import { defineConfig } from "vite";
-import tsconfigPaths from "vite-tsconfig-paths";
 
 export default defineConfig({
-  plugins: [reactRouter(), tsconfigPaths()],
+  resolve: { tsconfigPaths: true },
+  plugins: [reactRouter()],
 });